% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertPower.R
\name{ConvertPower}
\alias{ConvertPower}
\title{ConvertPower}
\usage{
ConvertPower(data, tvgCFac = 2)
}
\arguments{
\item{data}{from ReadEKRaw function (must run \code{\link{ReadEKRaw}} first). A list of three objects called data,
(1) File header, (2) pings, (3) GPS data. These three objects are themselves lists.}

\item{tvgCFac}{Time Varied Gain correction factor; default = 2; specify based upon application for the ES/K 60 the recommended values to apply are as follows:
for Sv data for echo integration = 2
TS data for use with method 1 single target detection operators (see Notes at URL) = 2
TS data for use with method 2 single target detection operators (see Notes at URL) = 0
http://support.echoview.com/WebHelp/Reference/Algorithms/Echosounder/Simrad/Simrad_Time_Varied_Gain_Range_Correction.htm}
}
\value{
adds a vector named Sv to each object in \code{data$pings} in (db re. 1 m^-1). This can then be used for further analysis such as with \code{\link{Echogram}}
}
\description{
Convert raw power to Sv
}
