% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateMatrix.R
\name{CreateMatrix}
\alias{CreateMatrix}
\title{CreateMatrix}
\usage{
CreateMatrix(data, nf = 0, use.bottom.line = FALSE,
  bottom.line.offset = 0, threshold = -30)
}
\arguments{
\item{data}{must be a data object that comes out of ReadEKRaw(fname) & ReadEKRaw_ConvertPower(data).
If matrix is to be constrained above the lake bottom then the data object must have also been processed by the FindBottom2(data).}

\item{nf}{nearfield distance, offset below surface in meters}

\item{use.bottom.line}{Use the lake bottom to bottom depth of the matrix; default is FALSE}

\item{bottom.line.offset}{depth offset from bottom in meters, default is 0 (can also be a negative number)}

\item{threshold}{data with backscatter (dB) stronger than this value are turned to NA's. Use this to remove possible scatter from fish}
}
\value{
data matrix of backscatter for each ping at each depth interval
}
\description{
Create data matrix for plotting and analysis.
}
\examples{
z = CreateMatrix(data, nf = 0.5, use.bottom.line = TRUE, bottom.line.offset = 0, threshold = -30)
}
