% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistTransect.R
\name{DistTransect}
\alias{DistTransect}
\title{DistTransect}
\usage{
DistTransect(data)
}
\arguments{
\item{data}{data object from that comes out of \code{\link{ReadEKRaw}}, \code{\link{ConvertPower}} and \code{\link{SubsetData}} The start of the transect needs to be the first point in the data for this to work, so \code{\link{SubsetData}} must have already been used.}
}
\value{
creates a new vector to data list named data$dist.alongtrans that contains distance from beginning of transect for each ping. Within each ping an element called GPS$dist.alongtrans is also created that indicated the distance along the transect for that specific ping.
}
\description{
or each ping in data file estimates distance along transect. Note that this is different as distance from shore.
}
