% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echogram.R
\name{Echogram}
\alias{Echogram}
\title{Echogram}
\usage{
Echogram(data, nf = 1, use.bottom.line = TRUE, bottom.line.offset = 1,
  threshold = -30, x = seq(1, length(data$pings), by = 1),
  y = data$pings[[2]]$rangeCorrected, ylim = rev(c(0, max(data$bottom.line +
  bottom.line.offset))), zstart = -70, zend = -30, xlim = range(x, finite
  = TRUE), zlim = c(zstart, zend), filename, height = 5, width = 7,
  ppi = 300, nlevels = 15, levels = pretty(zlim, nlevels),
  plot.title = "", key.title = "Sv", xlab = " Index", asp = NA,
  xaxs = "i", yaxs = "i", las = 1)
}
\arguments{
\item{data}{object from \code{\link{ConvertPower}} or \code{\link{FindBottom}}}

\item{nf}{nearfield distance, offset below surface in meters to be excluded from plot; default is 1}

\item{use.bottom.line}{default is TRUE, when set to default only backscatter from the bottom line offset to the nf range are plotted.}

\item{bottom.line.offset}{value to specify depth offset in meters below (positive values) or above (negative values) the bottom line that are to be plotted; default is 1 m.}

\item{threshold}{data with backscatter (dB) stronger than this value are excluded. Use this to remove possible scatter from fish; default is -30 dB.}

\item{x}{vector of values to be used for the x-axis. Must be the same length as number of pings; default is a vector 1:length(pings).}

\item{y}{vector of depth intervals form sample data; default is data$pings[[2]]$rangeCorrected. THIS SHOULD NOT BE CHANGED.}

\item{ylim}{range of y values; default is rev(c(0, max(data$bottom.line + bottom.line.offset))), THIS SHOUKD NOT BE CHANGED.}

\item{zstart}{lowest value of z (backscatter); default is -70 dB}

\item{zend}{highest value of z (backscatter); default is -30 dB}

\item{xlim}{range of x values; default is range(x, finite=TRUE). THIS SHOULD NOT BE CHANGED.}

\item{zlim}{range of z values base on zstart and zend, THIS SHOULD NOT BE CHANGED.}

\item{filename}{file name in the form “filename.png”; must be specified so the file can be written}

\item{height}{value in inches that specifies the height of the image; default is 5}

\item{width}{value in inches that specifies the width of the image; default is 7}

\item{ppi}{resolution of image; default is 300, make smaller for smaller image files}

\item{nlevels}{number of levels for color scale; default is 15}

\item{levels}{pretty separations for colors of z and scale bare. THIS SHOULD NOT BE CHANGED}

\item{plot.title}{character vector in quotes to label plot; default is “”}

\item{key.title}{character vector in quotes of title of color key; default is “Sv"}

\item{xlab}{character vector in quotes of title of color key; default is  “Index"}

\item{asp}{aspect ratio y/x, default NA. See \code{\link[graphics]{plot.window}} for more information.}

\item{xaxs}{x axis interval calculation style. Default "i", see \code{\link[graphics]{par}} for more information.}

\item{yaxs}{x axis interval calculation style. Default "i", see \code{\link[graphics]{par}} for more information.}

\item{las}{axis label style. See \code{\link[graphics]{par}}}
}
\description{
creates an echogram with a scale bar
}
\examples{
Echogram.pixel(data, filename = "RPpixelfunction.png",plot.title = "Rock Pond", threshold = -45, use.bottom.line = TRUE)
}
