% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBottom.R
\name{FindBottom}
\alias{FindBottom}
\title{FindBottom}
\usage{
FindBottom(data, nf = 1, d = -30, smooth = TRUE)
}
\arguments{
\item{data}{A ReadEKRaw_ConvertPower data object (i.e., you must run ReadEKRaw and then ConvertPower first). A list of three objects called data, (1) File header, (2) pings, (3) GPS data. These three objects are themselves lists.}

\item{nf}{near field range in meters, user specified, defines the depth range from 0 - nf to exclude; defaults is 1 m}

\item{d}{discrimination level in dB, pixels below this value are not considered when finding the bottom; default is -30 dB. Echoview default is -50 dB}

\item{smooth}{defaults to TRUE}
}
\value{
adds a vector named bot.depth to each object in data$pings. Also generates a new element to data called bottom.line  that is a vector of the bottom depths at all indexed value in 1:length(data$pings), this vector is created for ease when plotting or other analyses.
}
\description{
Find bottom in backscatter matrix.
}
