% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadEKRaw.R
\name{ReadEKRaw}
\alias{ReadEKRaw}
\title{ReadEKRaw}
\usage{
ReadEKRaw(fname)
}
\arguments{
\item{fname}{character path to a .raw file from Simrad Ex60 transceiver}
}
\value{
list containing 3 elements:

\tabular{ll}{
header \tab contains the header information for the file, and transceiver(s) including transceiver configuration and settings. \cr
pings \tab contains data from each ping as a list \cr
GPS \tab GPS data for each ping in .raw file (see
}
}
\description{
parse the binary .raw files obtained from a Simrad Ex60 transceiver
}
