% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{SubsetData}
\alias{SubsetData}
\title{SubsetData}
\usage{
SubsetData(data, start, end, method)
}
\arguments{
\item{data}{a data object from ReadEKRaw, ConvertPower or FindBottom}

\item{start}{time as character in format:"YYYY-MM-DD HH:MM:SS GMT" or decimal degrees as a number with the form c(lat,long)}

\item{end}{see start}

\item{method}{specified by user as either "time" or "latlong", must match units in start and end}
}
\value{
individual transects from each lake that can then be plotted or further analyzed. Remember, that plotting function only works if the first ping in the data set is the start of the transect.
}
\description{
function to subset data based on user specified start and end time or by GPS lat/longs
}
